feat: create CodegenActions; other minor refactors
