+++ packages/graphql/schemas/schema.graphql
@@ -1583,9 +1583,6 @@ type Query {
 
 """Properties describing a React component"""
 type ReactComponentDescriptor {
-  """A description for the component"""
-  description: String!
-
   """The name of the component"""
   displayName: String!
 }
