+++ packages/graphql/src/schemaTypes/objectTypes/gql-ReactComponentDescriptor.ts
@@ -4,10 +4,6 @@ export const ReactComponentDescriptor = objectType({
   name: 'ReactComponentDescriptor',
   description: 'Properties describing a React component',
   definition (t) {
-    t.nonNull.string('description', {
-      description: 'A description for the component',
-    })
-
     t.nonNull.string('displayName', {
       description: 'The name of the component',
     })
